+++ packages/ubs/public/settings.js
@@ -1,25 +1,22 @@
-/**
- * Created with IntelliJ IDEA.
- * User: mystetskyy
- * Date: 20.12.13
- * Time: 18:26
- * To change this template use File | Settings | File Templates.
- */
+const UB = require('@unitybase/ub-pub')
+const $App = require('@unitybase/adminui-pub')
 
-UB.ns('UBS.Settings');
-(function() {
-    var data = {};
-    UB.Repository('ubs_settings')
-        .attrs(['ID', 'settingKey', 'name', 'description', 'type', 'settingValue', 'defaultValue'])
-        .selectAsObject()
-        .done(function(response){
-            data = response; //UBConnection.selectResultToArrayOfObjects(response);
-            UBS.Settings.findByKey = function(key){
-                var res = _.find(data, {settingKey: key});
-                if (res === undefined){
-                   UB.logDebug('value fom UBS.Settings['+key+'] not found');
-                }
-                return res;
-            };
-        })
-})();
+let __settingsData = []
+let Settings = UB.ns('UBS.Settings')
+Settings.findByKey = function (key) {
+  let res = __settingsData.find(elm => elm.settingKey === key)
+  if (res === undefined) {
+    UB.logDebug(`value fom UBS.Settings[${key}] not found`)
+  }
+  return res
+}
+
+$App.modelLoadedPromise = $App.modelLoadedPromise.then(() => {
+  return UB.Repository('ubs_settings')
+    .attrs(['ID', 'settingKey', 'name', 'description', 'type', 'settingValue', 'defaultValue'])
+    .selectAsObject()
+    .then(function (data) {
+      __settingsData = data
+      return true
+    })
+})